requirements.txt
ppb>=3.2
pillow==10.3.0